tournament-service/requirements.txt
fastapi
uvicorn
motor
pydantic
pyjwt